Add splid service
